# docs/api/apex-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryView": {
        "properties": {
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "environment_id"
        ],
        "title": "InventoryView",
        "type": "object"
      },
...
      "ServiceInfo": {
        "additionalProperties": false,
        "properties": {
          "image": {
            "default": "",
            "maxLength": 2048,
            "pattern": "^[^\\x00]*$",
            "title": "Image",
            "type": "string"
          },
          "name": {
            "maxLength": 253,
            "minLength": 1,
            "pattern": "^[^\\x00]*$",
            "title": "Name",
            "type": "string"
          },
          "replicas": {
            "default": 1,
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Replicas",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "ServiceInfo",
        "type": "object"
      },
...
      "SnapshotView": {
        "properties": {
          "scanned_at": {
            "format": "date-time",
            "title": "Scanned At",
            "type": "string"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceInfo"
            },
            "title": "Services",
            "type": "array"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "scanned_at",
          "services",
          "stale"
        ],
        "title": "SnapshotView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "APEX Orchestration Engine \u2014 Domain API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/inventory/environments/{environment_id}": {
      "get": {
        "operationId": "getEnvironmentInventory",
        "parameters": [
          {
            "description": "Environment id",
            "in": "path",
            "name": "environment_id",
            "required": true,
            "schema": {
              "description": "Environment id",
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[^\\x00]*$",
              "title": "Environment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Environment Inventory",
        "tags": [
          "inventory"
        ]
      }
    },
    "/v1/inventory/environments/{environment_id}/rescan": {
      "post": {
        "operationId": "rescanEnvironment",
        "parameters": [
          {
            "description": "Environment id",
            "in": "path",
            "name": "environment_id",
            "required": true,
            "schema": {
              "description": "Environment id",
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[^\\x00]*$",
              "title": "Environment Id",
              "type": "string"
            }
          },
          {
            "description": "Explicit cluster-inventory connection id",
            "in": "query",
            "name": "connection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "minLength": 1,
                  "pattern": "^[^\\x00]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Explicit cluster-inventory connection id",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rescan Environment",
        "tags": [
          "inventory"
        ]
      }
    },
...
  }
}